Extract writable object
